# Conflicts:
#	CMakeLists.txt
#	Src/Core/Definitions.h
#	Src/Tools/stb_include.cpp
#	Src/imgui/SideBar.h

+++ Src/CRLCamera/ThreadPool.h
@@ -87,7 +87,7 @@ namespace VkRender {
             return res;
         }
 
-        uint32_t getTaskListSize() {
+        size_t getTaskListSize() {
             return tasks_.size();
         }
